fix(packages/components/button): fix default background color
The default value of the `--btn-background` CSS variable is now
`rgb(var(--arc-color-default))` instead of `inital` and the
`.button--filled.button--default` CCS selector now does not apply a
value to the `background-color` property. This allows the button to
be set be the `--btn-background` CSS variable by defau by default.

BREAKING CHANHGE: To enable this change the `ArcButton.color` default property value has
been changed from `primary` to `default`. This is a breaking change for
anyone using the `ArcButton` component with the `color` property not
set.