feat(accessibility): add slot to allow for extra options controls (#211)
This adds a slot with the name of options to the `ArcAccessibility` component. This allows for
the addition of extra controls to be added to the accessibility component.

It is required to listen to the `arc-show-accessibility` event on the `ArcNavbar` component and
call `ArcAccessibility.show()` to show the accessibility component.

It is also required to implment all the functionality for the extra controls placed within the
slot. Only the default controls are implemented in the `ArcAccessibility` component.